Implemented deisgn patterns of the adapter, filter and notifier.

+++ cms/PVPoints.php
@@ -44,7 +44,12 @@ class PVPoints extends PVStaticObject {
 	 * @access public
 	 */
 	public static function addPoint($args=array()){
+		
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $args);
+		
 		$args += self::getPointsDefaults();
+		$args = self::_applyFilter( get_class(), __FUNCTION__ , $args);
 		$args=PVDatabase::makeSafe($args);
 		extract($args);
 		
@@ -56,7 +61,8 @@ class PVPoints extends PVStaticObject {
 			
 		$query="INSERT INTO ".PVDatabase::getPointsTableName()."(user_id, content_id, comment_id, app_id, point_value, point_type, user_ip, point_date) VALUES( '$user_id', '$content_id', '$comment_id', '$app_id', '$point_value', '$point_type', '$user_ip', '$point_date' ) ";
 		$point_id=PVDatabase::return_last_insert_query($query, 'point_id', PVDatabase::getPointsTableName());
-			
+		$this->_notify(get_class().'::'.__FUNCTION__, $point_id, $args);
+		
 		return $point_id;
 	}//end addUserPoster
 	
@@ -77,7 +83,12 @@ class PVPoints extends PVStaticObject {
 	 * @access public
 	 */
 	public static function addUniquePoint($args=array() ){
+		
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $args);
+		
 		$args += self::getPointsDefaults();
+		$args = self::_applyFilter( get_class(), __FUNCTION__ , $args);
 		$args=PVDatabase::makeSafe($args);
 		extract($args);
 		
@@ -116,7 +127,9 @@ class PVPoints extends PVStaticObject {
 				
 		if($allow_insert==true){	
 			$query="INSERT INTO ".PVDatabase::getPointsTableName()."(user_id, content_id, comment_id, app_id, point_value, point_type, user_ip, point_date)  VALUES( '$user_id', '$content_id', '$comment_id', '$app_id', '$point_value', '$point_type', '$user_ip', '$point_date' ) ";	
-			return $point_id=PVDatabase::return_last_insert_query($query, 'point_id', PVDatabase::getPointsTableName());
+			$point_id=PVDatabase::return_last_insert_query($query, 'point_id', PVDatabase::getPointsTableName());
+			$this->_notify(get_class().'::'.__FUNCTION__, $point_id, $args);
+			return $point_id;
 		}
 		return FALSE;
 	}//end addUserPoster
@@ -145,8 +158,14 @@ class PVPoints extends PVStaticObject {
 	 * @access public
 	 */
 	public static function getPointsList($args=array()){
+		
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $args);
+		
 		$args += self::getPointsDefaults();
 		$args += self::_getSqlSearchDefaults();
+		$args = self::_applyFilter( get_class(), __FUNCTION__ , $args);
+		
 		$custom_where=$args['custom_where'];
 		$custom_join=$args['custom_join'];
 		$custom_select=$args['custom_select'];
@@ -403,6 +422,8 @@ class PVPoints extends PVStaticObject {
     	}//end while
     	
     	$content_array=PVDatabase::formatData($content_array);
+		$this->_notify(get_class().'::'.__FUNCTION__, $content_array, $args);
+		
     	return $content_array;
 	}//end getUserPointsList
 	
@@ -416,13 +437,18 @@ class PVPoints extends PVStaticObject {
 	 */
 	public static function getPoint($point_id){
 		
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $point_id);
+			
+			$point_id = self::_applyFilter( get_class(), __FUNCTION__ , $point_id);
+		
 		if(!empty($point_id)){
-			
+				
+			$point_id=PVDatabase::makeSafe($point_id);
 			$query="SELECT * FROM ".PVDatabase::getPointsTableName()." WHERE point_id='$point_id'";
-			
 			$result=PVDatabase::query($query);
-			
 			$row = PVDatabase::fetchArray($result);
+			$this->_notify(get_class().'::'.__FUNCTION__, $row);
 			
 			return $row;
 		}
@@ -446,7 +472,12 @@ class PVPoints extends PVStaticObject {
 	 * @access public
 	 */
 	public static function updatePoint($args=array()){
+			
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $args);
+		
 		$args += self::getPointsDefaults();
+		$args = self::_applyFilter( get_class(), __FUNCTION__ , $args);
 		$args=PVDatabase::makeSafe($args);
 		extract($args);
 		
@@ -456,7 +487,7 @@ class PVPoints extends PVStaticObject {
 		
 		$query="UPDATE ".PVDatabase::getPointsTableName()." SET point_value='$point_value', content_id='$content_id', comment_id='$comment_id', app_id='$app_id', point_type='$point_type', user_id='$user_id', point_date='$point_date' WHERE point_id='$point_id'";
 		PVDatabase::query($query);
-		
+		$this->_notify(get_class().'::'.__FUNCTION__, $args);
 	}//end updateUserPoint
 	
 	/**
@@ -469,11 +500,17 @@ class PVPoints extends PVStaticObject {
 	 */
 	public static function deletePoint($point_id){
 		
+		if(self::_hasAdapter(get_class(), __FUNCTION__) )
+			return self::_callAdapter(get_class(), __FUNCTION__, $point_id);
+		
+		$point_id = self::_applyFilter( get_class(), __FUNCTION__ , $point_id);
+		
 		if(!empty($point_id)){
 			$point_id=PVDatabase::makeSafe($point_id);
 			$query="DELETE FROM ".PVDatabase::getPointsTableName()." WHERE point_id='$point_id' ";
 			
 			PVDatabase::query($query);
+			$this->_notify(get_class().'::'.__FUNCTION__, $point_id);
 		}
 	}//end deleteUserPoint
